DRF/requirements.txt
asgiref==3.8.1
Django==5.2.1
django-auth-ldap==5.2.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.9.0
python-decouple==3.8
python-ldap==3.4.4
sqlparse==0.5.3
gunicorn==20.1.0
